adjust process logic of views.py.

+++ travelproject/bot/views.py
@@ -72,7 +72,9 @@ next_type_hash.update({'hotel_name_input' : 'instant'}) # update for BREAK POINT
 # acceptable type define
 message_accept_type = ['entering_message' ,
                        'sightseeing' ,
-                       'hotel_name_input']
+                       'hotel_name_input' ,
+                       'food_recommend']
+
 postback_accept_type = [
                         'admin_area' ,
                         'queried_date' ,
@@ -138,57 +140,95 @@ def handle_message(event):
 
     else:
         type_header = client_obj.type_header # find the empty client attributes , it's actually the type_header now!
-    msg = event.message.text # parse messages from event object
+
 
     print('DEBUG handle_message type header: ', type_header)
 
-    # 這邊處理一種誤傳 message (本來是 postback button)狀況
-    # check the type is acceptable to message event; if not , it's belong to postback event
-    if type_header in message_accept_type:
 
-        # if type == entering_message or sightseeing or hotel name input , transfer to return postback
-        if type_header in ['entering_message','sightseeing']:
+    msg = event.message.text # parse messages from event object
+    if '重新搜尋' in msg:
 
-            other_msg = greeting_message if type_header == 'entering_message' else {}
+        # in every type stage , every time you key in "重新搜尋" or "re-search" ,
+        # it will return to beginning of search
 
-            save_attr_to_database(type_header, client_obj, msg) # saving data to database if data_key exist in object attr
-            return_postback(event ,
-                            client_obj = client_obj ,
-                            type_header = type_header ,
-                            other_msg = other_msg)
+        '''
+         Not saving attrs here!
+        '''
+        type_header = type_header_backward(client_obj , target_type='entering_message')
+        other_msg = '請重新選擇您要去的縣市~'
+        return_postback(event,
+                        client_obj=client_obj,
+                        type_header=type_header,
+                        other_msg=other_msg)
 
-        elif type_header == 'hotel_name_input' :
+    else:
 
-            selected_hotel = get_similar_name_hotel(msg) # check if this hotel name exist in hotel list
+        # 這邊處理一種誤傳 message (本來是 postback button)狀況
+        # check the type is acceptable to message event; if not , it's belong to postback event
 
-            if selected_hotel:
-                save_attr_to_database(type_header , client_obj , selected_hotel.source_name)
+        if type_header in message_accept_type:
+
+            # if type == entering_message or sightseeing or hotel name input , transfer to return postback
+            if type_header in ['entering_message' , 'sightseeing']:
+
+                other_msg = greeting_message if type_header == 'entering_message' else ''
+
+                save_attr_to_database(type_header, client_obj, msg) # saving data to database if data_key exist in object attr
+                return_postback(event ,
+                                client_obj = client_obj ,
+                                type_header = type_header ,
+                                other_msg = other_msg)
+
+            # in self-preparing 'hotel_name_input' stage , check the input name exist or not
+            elif type_header == 'hotel_name_input' :
+
+                selected_hotel = get_similar_name_hotel(msg) # check if this hotel name exist in hotel list
+
+                if selected_hotel:
+
+                    # to return "instant" postback next
+                    save_attr_to_database(type_header , client_obj , selected_hotel.source_name)
+                    return_postback(event,
+                                    client_obj=client_obj,
+                                    type_header=type_header)
+                else:
+
+                    '''
+                     Not saving attrs here!
+                    '''
+                    type_header = type_header_backward(client_obj)
+                    other_msg = '找不到您輸入的飯店喔 ,請確認名稱 ~ '
+                    return_message(event,
+                                   client_obj=client_obj,
+                                   type_header=type_header,
+                                   other_msg = other_msg)
+
+            # in 'food_recommend' stage and want to return back to 'recommend'
+            elif type_header == 'food_recommend' and '返回推薦' in msg:
+
+                '''
+                 Not saving attrs here!
+                '''
+
+                type_header = type_header_backward(client_obj , target_type='sightseeing') # for food recommend stage , return to hotel recommend stage
                 return_postback(event,
                                 client_obj=client_obj,
                                 type_header=type_header)
-            else:
 
-                type_header = get_pre_type_header(type_header, client_obj)
-                other_msg = '找不到您輸入的飯店喔 ,請確認名稱 ~ '
-                return_message(event,
-                               client_obj=client_obj,
-                               type_header=type_header,
-                               other_msg = other_msg)
+
+            else:
+                save_attr_to_database(type_header, client_obj, msg)
+                return_message(event ,
+                               client_obj = client_obj ,
+                               type_header = type_header  )
 
         else:
-            save_attr_to_database(type_header, client_obj, msg)
-            return_message(event ,
-                           client_obj = client_obj ,
-                           type_header = type_header  )
-    else:
 
-        # TODO: 這邊有個 BUG XD , 假設現在是 'hotel_name_input' 後的 'instant' , 但這邊不小心輸入了 message , 就會造成 'instant' 倒回 'recommend'
-
-        # return type_header for "1 stage" , re-send postback event
-        type_header = get_pre_type_header(type_header, client_obj)
-        return_postback(event,
-                        client_obj=client_obj,
-                        type_header=type_header)
+            # return type_header for "1 stage" , re-send postback event
+            type_header = type_header_backward(client_obj)
+            return_postback(event,
+                            client_obj=client_obj,
+                            type_header=type_header)
 
 def return_message(event,
                    client_obj,
@@ -203,7 +243,6 @@ def return_message(event,
     # next_type_header : the type will return to client side
 
     '''
-    contents = None
 
     # special handle for BREAK POINT "NeedRecommendOrNot" ; if got "N" in 'NeedRecommendOrNot' stage , "fork" to 'hotel_name_input'
     if type_header == 'NeedRecommendOrNot' and getattr(client_obj , type_header) == 'N':
@@ -222,10 +261,9 @@ def return_message(event,
 
     client_obj.save()
 
+    print('DEBUG return_message client_obj.type_record ', client_obj.type_record)
 
-
-    print('DEBUG return_message type header and next_type_header: ', type_header , next_type_header)
-
+    contents = None
 
     if not next_type_header:
         raise ValueError('No next type exist!!!')
@@ -246,7 +284,8 @@ def return_message(event,
     reply_action = [TextSendMessage(text=contents)]
     if other_msg:
         for _ , msg in other_msg.items():
-            reply_action.insert(0, TextSendMessage(text=msg))  # insert other msg in front of button reply
+            if msg:
+                reply_action.insert(0, TextSendMessage(text=msg))  # insert other msg in front of button reply
 
     line_bot_api.reply_message(
         event.reply_token,
@@ -271,7 +310,7 @@ def handle_postback(event):
         client_obj = Line_client.create_obj_by_dict(user_id = event.source.user_id,
                                                     query_date=datetime.date.today())
 
-    # 這邊防止剛進來手殘按到之前的 button .
+    # TODO(怪怪?) : 這邊防止剛進來手殘按到之前的 button .
     if not client_obj.type_header and not client_obj.entering_message:
         client_obj.type_header = type_header = 'entering_message'
         client_obj.type_record.append('entering_message')
@@ -279,26 +318,68 @@ def handle_postback(event):
 
     else:
         # Got a post event back , and parse the current type and data
-        type_header = event.postback.data.split('&')[0]   # get type header from postback data ; NOTE THAT 這邊為了避免 postback to postback 誤觸情況
+        type_header = event.postback.data.split('&')[0]   # get type header from postback data ; NOTE THAT 這邊為了避免 postback to postback 誤觸情況 , 取 template type 來跟 client_obj.type_header 來做比對
         pre_postback_data = event.postback.params['date'] if type_header == 'queried_date' else event.postback.data.split('&')[1]  # get data from postback data
 
     # 這邊處理兩種 postback button (本來是 message or postback button) 誤觸狀況 :
     # Firstly, check the type header from postback is == from client.obj
     if type_header == client_obj.type_header:
 
+        # BREAK POINT for 'silence' and 'hotel_name_input'
         # if NeedRecommendOrNot == "N" or food type , "fork" to return message
-        if (type_header == 'NeedRecommendOrNot' and pre_postback_data == "N") or \
-            type_header == "recommend" and 'FoodRecommend' in pre_postback_data or \
-            type_header == "food":
+        if (type_header == 'NeedRecommendOrNot' and pre_postback_data == "N") or type_header == "food":
 
-            # 這邊如果是 'recommend' 下會有兩種狀況 ; 一種是要 recommend hotel ; 另一種是要 recommend food
-            # 這兩種都是在 'recommend' 儲存 hotel source_name ,
-            # 唯一差別是一種是 send postback , 一種是 send message
             save_attr_to_database(type_header , client_obj , pre_postback_data)
             return_message(event,
                            client_obj=client_obj,
                            type_header=type_header)
 
+
+        # BREAK POINT for "instant" and "food_recommend"
+        # if "FoodRecommend" in postback button data , "fork" to return message
+        elif type_header == "recommend" and 'food_recommend' in pre_postback_data :
+
+            '''
+             Not saving attrs here!
+            '''
+
+            other_msg = '請輸入 "返回推薦" 以回到飯店推薦頁面~'
+            return_message(event ,
+                           client_obj=client_obj ,
+                           type_header=type_header ,
+                           other_msg = other_msg)
+
+        # If it goes to the 'leaf' of the selection process ,
+        # choose to return to 'more recommend' or 'beginning of search'
+        elif type_header == 'instant':
+
+            '''
+             Not saving attrs here!
+            '''
+
+            # for returning to recommend list
+            if 'return_recommend' in pre_postback_data:
+
+                # if it's not recommend mode originally , keep going collecting user data.
+                if 'recommend' not in client_obj.type_record:
+                    type_header = type_header_backward(client_obj, target_type='num_rooms')
+                else:
+                    type_header = type_header_backward(client_obj, target_type='sightseeing')
+                other_msg = ''
+
+            # for returning to beginning of search
+            elif 'return_search' in pre_postback_data:
+
+                type_header = type_header_backward(client_obj, target_type='entering_message')
+                other_msg = '請重新選擇您要去的縣市~'
+
+
+            return_postback(event,
+                            client_obj=client_obj,
+                            type_header=type_header,
+                            other_msg=other_msg)
+
+        # else , directly return postback .
         else:
             save_attr_to_database(type_header, client_obj, pre_postback_data)
             return_postback(event,
@@ -312,8 +393,7 @@ def handle_postback(event):
         # Secondly , if not the same type between type header from postback and from client.obj
         # return type_header for "1 stage" , re-send postback(or message) event
 
-        type_header = client_obj.type_header # let type_header equal to client_obj.type_header if these 2 are not the same type at this time
-        type_header = get_pre_type_header(type_header, client_obj)
+        type_header = type_header_backward(client_obj)
 
         # And then to check what type actually it is.
         if type_header in postback_accept_type:
@@ -342,7 +422,8 @@ def return_postback(event,
     client_obj.type_record.append(next_type_hash.get(type_header, None))
     client_obj.save() # save the update on type_header
 
-    print('DEBUG return_postback type header and next_type_header: ', type_header ,  next_type_header)
+
+    print('DEBUG return_postback client_obj.type_record ', client_obj.type_record)
 
     if not next_type_header:
         raise ValueError('No next type exist!!!')
@@ -387,8 +468,9 @@ def return_postback(event,
 
     reply_action = [FlexSendMessage(alt_text='FlexTemplate',contents=contents)]
     if other_msg:
-        for _, msg in other_msg.items():
-            reply_action.insert(0 , TextSendMessage(text=msg)) # insert other msg in front of button reply
+        for _ , msg in other_msg.items():
+            if msg:
+                reply_action.insert(0 , TextSendMessage(text=msg)) # insert other msg in front of button reply
 
     line_bot_api.reply_message(
         event.reply_token,
@@ -445,6 +527,8 @@ def get_hotel_instance(client_obj):
 
     dict_list = []
     for ins_obj in instant_objs:
+
+
         ins_dict = ins_obj.__dict__
         ins_dict.update({'pic_link': pic_link})
         dict_list.append(ins_dict)
@@ -489,18 +573,35 @@ def get_similar_name_hotel(selected_hotel):
     return selected_hotel
 
 
-def get_pre_type_header(type_header , client_obj):
+def type_header_backward(client_obj , target_type = None):
 
     '''
     # function: back to the previous stage of type
 
-    :param type_header:  the stage of type now client in
     :param client_obj:  the client object
+    :param target_type : the type stage want to return
     :return: the previous stage of type
     '''
 
-    client_obj.type_record.pop() # remove current type
-    type_header = client_obj.type_header = client_obj.type_record[-1] # get last element of type record (previous type) after rm cur type as cur type
-    client_obj.save() # save change
+    # if not assign type header , default backward 1 stage to previous type
+    if not target_type:
+
+        client_obj.type_record.pop()  # remove current type
+        type_header = client_obj.type_header = client_obj.type_record[-1]  # get last element of type record (previous type) after rm cur type as cur type
+        client_obj.save()  # save change
+
+    # if assign type , back to that type stage
+    else:
+
+        type_record = client_obj.type_record
+        if target_type not in type_record:
+            raise ValueError('Target type not in type record !')
+
+        target_idx = type_record.index(target_type)
+        type_record = type_record[:target_idx+1] # remove type after target type
+        type_header = client_obj.type_header = type_record[-1] # get now type header
+
+        client_obj.type_record = type_record
+        client_obj.save() # save change
 
     return type_header
